requirements.txt
gdown==4.4.0
matplotlib==3.5.1
numpy==1.22.3
opencv_python==4.5.5.64
opencv_contrib_python==4.5.5.64
Pillow==9.1.0
pycocotools==2.0.4
torch==1.10.2
torchvision==0.11.3
